requirements.txt
pypinyin
cffi==1.14.0
cryptography==2.8
lxml==4.5.0
pycparser==2.20
PyMuPDF==1.16.16
PyMySQL==0.9.3
python-docx==0.8.10
six==1.14.0
xlrd==1.2.0
